+++ src/react/App.js
@@ -2,9 +2,9 @@ import { Route, Routes } from 'react-router-dom'
 
 import NavigationBar from './components/navbar/navbar'
 import Footer from './components/footer/footer'
-import Home from './components/pages/home/home'
-import Register from './components/pages/register/register'
-import Contact from './components/pages/contact/contact'
+import Home from './pages/home/home'
+import Register from './pages/register/register'
+import Contact from './pages/contact/contact'
 
 export default function App() {
 	return (
